Change parameter types for matrix functionso
troquei as funçoes float por int, o float tava dando erro de proporçao e o correto era dar um valor inteiro

+++ Headers/funcoesMatriz.h
@@ -5,7 +5,7 @@
 
 void gerarMat(int mat[MAX][MAX]);
 int calcDet(int mat[MAX][MAX]);
-void calcMatAdj(int mat[MAX][MAX], float matAdj[MAX][MAX]);
-void calcMatInv(float matAdj[MAX][MAX], float matInv[MAX][MAX], int det);
+void calcMatAdj(int mat[MAX][MAX], int matAdj[MAX][MAX]);
+void calcMatInv(int matAdj[MAX][MAX], int matInv[MAX][MAX], int det);
 
 #endif // FUNCOESMATRIZ_H
